Fix several warnings in gio/tests/converter-stream.c
gio/tests/converter-stream.c: In function ‘g_expander_converter_convert’:
gio/tests/converter-stream.c:128:21: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
       for (i = 0; i < block_size; i++)
                     ^
gio/tests/converter-stream.c: In function ‘g_compressor_converter_convert’:
gio/tests/converter-stream.c:234:23: error: comparison of integer expressions of different signedness: ‘long int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
       if (in_end - in < block_size)
                       ^
gio/tests/converter-stream.c:244:21: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
       for (i = 0; i < block_size; i++)
                     ^
gio/tests/converter-stream.c:257:33: error: comparison of integer expressions of different signedness: ‘long int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
       if (v == 0 && in_end - in == block_size && (flags & G_CONVERTER_INPUT_AT_END) == 0)
                                 ^~
gio/tests/converter-stream.c: In function ‘test_expander’:
gio/tests/converter-stream.c:356:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < sizeof(unexpanded_data); i++)
                 ^
gio/tests/converter-stream.c: In function ‘test_compressor’:
gio/tests/converter-stream.c:445:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   for (i = 0; i < expanded_size; i++)
                 ^
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from gio/tests/converter-stream.c:23:
gio/tests/converter-stream.c:454:16: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
    g_assert (i == expanded_size -1);
                ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
gio/tests/converter-stream.c:454:4: note: in expansion of macro ‘g_assert’
    g_assert (i == expanded_size -1);
    ^~~~~~~~
gio/tests/converter-stream.c: In function ‘test_converter_pollable’:
gio/tests/converter-stream.c:1077:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   for (i = 0; i < expanded_size; i++)
                 ^
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from gio/tests/converter-stream.c:23:
gio/tests/converter-stream.c:1086:16: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
    g_assert (i == expanded_size -1);
                ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
gio/tests/converter-stream.c:1086:4: note: in expansion of macro ‘g_assert’
    g_assert (i == expanded_size -1);
    ^~~~~~~~
gio/tests/converter-stream.c: In function ‘main’:
gio/tests/converter-stream.c:1220:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (compressor_tests); i++)
                 ^
gio/tests/converter-stream.c:1223:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (truncation_tests); i++)
                 ^
gio/tests/converter-stream.c:1226:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (charset_tests); i++)
                 ^

+++ gio/tests/converter-stream.c
@@ -94,7 +94,7 @@ g_expander_converter_convert (GConverter *converter,
 {
   const guint8 *in, *in_end;
   guint8 v, *out;
-  int i;
+  gsize i;
   gsize block_size;
 
   in = inbuf;
@@ -210,7 +210,7 @@ g_compressor_converter_convert (GConverter *converter,
 {
   const guint8 *in, *in_end;
   guint8 v, *out;
-  int i;
+  gsize i;
   gsize block_size;
 
   in = inbuf;
@@ -231,7 +231,7 @@ g_compressor_converter_convert (GConverter *converter,
 	block_size = v * 1000;
 
       /* Not enough data */
-      if (in_end - in < block_size)
+      if (((gsize) (in_end - in)) < block_size)
 	{
 	  if (*bytes_read > 0)
 	    break;
@@ -254,7 +254,8 @@ g_compressor_converter_convert (GConverter *converter,
 	    }
 	}
 
-      if (v == 0 && in_end - in == block_size && (flags & G_CONVERTER_INPUT_AT_END) == 0)
+      if (v == 0 && ((gsize) (in_end - in)) == block_size &&
+          (flags & G_CONVERTER_INPUT_AT_END) == 0)
 	{
 	  if (*bytes_read > 0)
 	    break;
@@ -297,7 +298,7 @@ test_expander (void)
   GConverter *expander;
   GConverter *converter;
   GError *error;
-  int i;
+  gsize i;
 
   expander = g_expander_converter_new ();
 
@@ -393,7 +394,7 @@ test_compressor (void)
   GOutputStream *mem_out, *cstream_out;
   GConverter *expander, *compressor;
   GError *error;
-  int i;
+  gsize i;
 
   expander = g_expander_converter_new ();
   expanded = g_malloc (100*1000); /* Large enough */
@@ -972,7 +973,7 @@ test_converter_pollable (void)
   GPollableOutputStream *pollable_out;
   GConverter *expander, *compressor;
   GError *error;
-  int i;
+  gsize i;
 
   expander = g_expander_converter_new ();
   expanded = g_malloc (100*1000); /* Large enough */
@@ -1207,7 +1208,7 @@ main (int   argc,
     { "/converter-input-stream/charset/fallbacks", "UTF-8", "Some characters just don't fit into latin1: πא", "ISO-8859-1", "Some characters just don't fit into latin1: \\CF\\80\\D7\\90", 4 },
   };
 
-  gint i;
+  gsize i;
 
   g_test_init (&argc, &argv, NULL);
